overhaul props and proptypes
